m2sdr_sata: split low-level helpers
Move CSR access, routing, streamer control, host-buffer access, pattern helpers and Etherbone bulk probing into m2sdr_sata_lowlevel.c/.h.

Keep the CLI command dispatch, catalog and workflow logic in m2sdr_sata.c so this step is a mechanical ownership split with no command-line changes.

Validated with make m2sdr_sata and ./m2sdr_sata --help.
